backend/requirements.txt
fastapi
uvicorn[standard]
pillow
gradio_client
python-multipart
requests
